coc: suggest rewording of reverse-ism clause
A while back, a conversation about this CoC came up over at https://github.com/videojs/video.js/pull/4160 when a PR for it was made. I spent some time thinking about it, and I'm wondering if it *is* worth rewording these bits. I'm hoping for some discussion and feedback on whether this is worth changing, whether the new edit is good, or whether the entire premise of changing it is basically preposterous: I don't want to neglect this particular aspect from this CoC, which is _meant_ to be thorough so that everyone knows the ground rules.